requirements.txt
streamlit
tensorflow
keras
numpy
pillow
opencv-python